+++ djangocms_text_ckeditor/__init__.py
@@ -1,3 +1,3 @@
 # -*- coding: utf-8 -*-
-__version__ = "3.2.1rc5"
+__version__ = "3.2.1"
 default_app_config = 'djangocms_text_ckeditor.apps.TextCkeditorConfig'