+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ CHANGELOG
 * Changed the way ckeditor js is loaded, everything is now loaded in one bundle
 * Changed the way ckeditor loads must-have external plugins, so user config
   would have no effect on the settings
+* Fixed error which prevented plugins extending TextPlugin to work
 * Fixed a regression where CKEditor would create unnecessary ``<p>`` tags around
   CMS plugins